playsafe-backend--main/age18OrNot/requirements.txt
streamlit
opencv-python
numpy
google-generativeai
python-dotenv
